Updated coordinates from int to float

+++ erland-game-tileadventure/trunk/src/erland/game/tileadventure/AnimationEditorLevelManager.java
@@ -104,7 +104,7 @@ public class AnimationEditorLevelManager extends LevelManager {
                 obj.setBlockType(image+1);
                 obj.setImage(prefix+"ground.gif", obj.getBlockType()-1);
                 obj.setPos((i)%map.getSizeX(), (i)/map.getSizeX(), 0);
-                map.setObject(obj, obj.getPosX(), obj.getPosY(), obj.getPosZ());
+                map.setObject(obj, (int)obj.getPosX(), (int)obj.getPosY(), (int)obj.getPosZ());
             }
             return new LevelInfo(map,info);
         }else {
